+++ app/ide-desktop/lib/dashboard/src/hooks/useLazyMemoHooks.ts
@@ -11,7 +11,7 @@ const UNSET_VALUE = Symbol('unset')
 /**
  * A hook that returns a memoized function that will only be called once
  */
-export function useLazyMemo<T>(factory: T | (() => T), deps: React.DependencyList): () => T {
+export function useLazyMemoHooks<T>(factory: T | (() => T), deps: React.DependencyList): () => T {
   return React.useMemo(() => {
     let cachedValue: T | typeof UNSET_VALUE = UNSET_VALUE
